Update: Used Int type that are OS-independent

+++ data_structures/stack/dynamic_stack.c
@@ -2,24 +2,29 @@
  * @file
  *
  * @brief
- * Dynamic [Stack](https://en.wikipedia.org/wiki/Stack_(abstract_data_type)), just like Dynamic Array, is a stack data structure whose the length or capacity (maximum number of elements that can be stored) increases or decreases in real time based on the operations (like insertion or deletion) performed on it.
+ * Dynamic [Stack](https://en.wikipedia.org/wiki/Stack_(abstract_data_type)),
+ * just like Dynamic Array, is a stack data structure whose the length or
+ * capacity (maximum number of elements that can be stored) increases or
+ * decreases in real time based on the operations (like insertion or deletion)
+ * performed on it.
  *
- * In this implementation, functions such as PUSH, POP, PEEK, show_capacity, isempty, and stack_size are coded to implement dynamic stack.
+ * In this implementation, functions such as PUSH, POP, PEEK, show_capacity,
+ * isempty, and stack_size are coded to implement dynamic stack.
  *
  * @author [SahilK-027](https://github.com/SahilK-027)
  *
  */
-#include <assert.h> /// to verify assumptions made by the program and print a diagnostic message if this assumption is false.
-#include <stdio.h>  /// for IO operations
-#include <stdlib.h> /// for including functions involving memory allocation such as `malloc`
-
+#include <assert.h>  /// to verify assumptions made by the program and print a diagnostic message if this assumption is false.
+#include <inttypes.h>  /// to provide a set of integer types with universally consistent definitions that are operating system-independent
+#include <stdio.h>     /// for IO operations
+#include <stdlib.h>  /// for including functions involving memory allocation such as `malloc`
 /**
  * @brief DArrayStack Structure of stack.
  */
 typedef struct DArrayStack
 {
-    int capacity, top; ///< to store capacity and top of the stack
-    int *arrPtr;       ///< array pointer
+    int capacity, top;  ///< to store capacity and top of the stack
+    int *arrPtr;        ///< array pointer
 } DArrayStack;
 
 /**
@@ -40,7 +45,8 @@ DArrayStack *create_stack(int cap)
 }
 
 /**
- * @brief As this is stack implementation using dynamic array this function will expand the size of the stack by twice as soon as the stack is full.
+ * @brief As this is stack implementation using dynamic array this function will
+ * expand the size of the stack by twice as soon as the stack is full.
  *
  * @param ptr Stack pointer
  * @param cap Capacity of stack
@@ -62,7 +68,9 @@ DArrayStack *double_array(DArrayStack *ptr, int cap)
 }
 
 /**
- * @brief As this is stack implementation using dynamic array this function will shrink the size of stack by twice as soon as the stack's capacity and size has significant difference.
+ * @brief As this is stack implementation using dynamic array this function will
+ * shrink the size of stack by twice as soon as the stack's capacity and size
+ * has significant difference.
  *
  * @param ptr Stack pointer
  * @param cap Capacity of stack
@@ -135,7 +143,8 @@ int pop(DArrayStack *ptr)
 }
 
 /**
- * @brief To retrieve or fetch the first element of the Stack or the element present at the top of the Stack.
+ * @brief To retrieve or fetch the first element of the Stack or the element
+ * present at the top of the Stack.
  *
  * @param ptr Stack pointer
  * @return int Top of the stack
@@ -156,13 +165,11 @@ int peek(DArrayStack *ptr)
  * @param ptr Stack pointer
  * @return int Current capacity of the stack
  */
-int show_capacity(DArrayStack *ptr)
-{
-    return ptr->capacity;
-}
+int show_capacity(DArrayStack *ptr) { return ptr->capacity; }
 
 /**
- * @brief The function is used to check whether the stack is empty or not and return true or false accordingly.
+ * @brief The function is used to check whether the stack is empty or not and
+ * return true or false accordingly.
  *
  * @param ptr Stack pointer
  * @return int returns 1 -> true OR returns 0 -> false
@@ -177,15 +184,13 @@ int isempty(DArrayStack *ptr)
 }
 
 /**
- * @brief Used to get the size of the Stack or the number of elements present in the Stack.
+ * @brief Used to get the size of the Stack or the number of elements present in
+ * the Stack.
  *
  * @param ptr Stack pointer
  * @return int size of stack
  */
-int stack_size(DArrayStack *ptr)
-{
-    return ptr->top + 1;
-}
+int stack_size(DArrayStack *ptr) { return ptr->top + 1; }
 
 /**
  * @brief Self-test implementations
@@ -196,7 +201,7 @@ static void test()
     DArrayStack *NewStack;
     int capacity = 1;
     NewStack = create_stack(capacity);
-    unsigned int arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
+    uint64_t arr[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
 
     printf("\nTesting Empty stack: ");
     assert(stack_size(NewStack) == 0);
@@ -207,7 +212,8 @@ static void test()
     for (int i = 0; i < 12; ++i)
     {
         int topVal = push(NewStack, arr[i]);
-        printf("Size: %d, Capacity: %d\n\n",stack_size(NewStack), show_capacity(NewStack));
+        printf("Size: %d, Capacity: %d\n\n", stack_size(NewStack),
+               show_capacity(NewStack));
         assert(topVal == i);
         assert(peek(NewStack) == arr[i]);
         assert(stack_size(NewStack) == i + 1);
@@ -239,6 +245,6 @@ static void test()
  */
 int main()
 {
-    test(); // run self-test implementations
+    test();  // run self-test implementations
     return 0;
 }
